Added the ability to include tasks without due dates (which closes #16). Also, fixed up the rule and completion behaviors to work as expected. This change required moving "Rule" to package "core," which is good. Finally, renamed "DynamicRuleParser" to "DynamicBehaviorEngine."
